+++ db/migrate/20230430185610_make_genpro_gov_sk_prosecutors_list_optional_for_appointments.rb
@@ -1,9 +1,9 @@
 class MakeGenproGovSkProsecutorsListOptionalForAppointments < ActiveRecord::Migration[6.0]
   def up
-    change_column :appointments, :genpro_gov_sk_prosecutors_list_id, :integer, null: true
+    change_column :appointments, :genpro_gov_sk_prosecutors_list_id, :bigint, null: true
   end
 
   def down
-    change_column :appointments, :genpro_gov_sk_prosecutors_list_id, :integer, null: false
+    change_column :appointments, :genpro_gov_sk_prosecutors_list_id, :bigint, null: false
   end
 end
